+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 ### Added
 
 - Added `message.Switch()`, `Map()` and `TryMap()`.
+- Added `message.Kind` enumeration.
 
 ### Removed